% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fibonacci_sum}
\alias{fibonacci_sum}
\title{Sum of Fibonacci Terms, Excluding the First}
\usage{
fibonacci_sum(n)
}
\arguments{
\item{n}{how many terms to sum}
}
\value{
the sum of the 2:n Fibonacci terms
}
\description{
Sum of Fibonacci Terms, Excluding the First
}
\details{
This calculates the number of elements arranged in a triangle with
one at the point, and each row containing \eqn{F_n} elements where
\eqn{F_n} is the nth Fibonacci number, calculated using
\code{\link{fibonacci_binet}}.
}
\examples{
fibonacci_sum(4)
# 11
}
