% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_frequencies.R
\docType{data}
\name{tidy_letters}
\alias{tidy_letters}
\title{Relative Proportions of Usage of a-z in English, French, German, and Spanish.}
\format{A \code{data.frame} with 104 rows and 3 variables:
\describe{
\item{Letter}{letter of the alphabet, a-z}
\item{Lang}{languge in which the letter is used}
\item{Prop}{relative proportion of use of the letter}
}}
\source{
\url{https://en.wikipedia.org/wiki/Letter_frequency}
}
\usage{
tidy_letters
}
\description{
A tidied dataset comparing the relative frequencies of letters from four languages.
}
\keyword{datasets}
