% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_plots.R
\name{user_plot}
\alias{user_plot}
\alias{ggshape_lang_diffs}
\alias{ggshape_english_usage}
\alias{ggshape_random_walk}
\alias{ggshape_rainbow_text}
\title{Example \code{user_plot} functions.}
\usage{
user_plot(x)

ggshape_lang_diffs(letter)

ggshape_english_usage(letter)

ggshape_random_walk(letter)

ggshape_rainbow_text(text)
}
\arguments{
\item{x, letter, text}{facetting variable}
}
\description{
Example \code{user_plot} functions.
}
